boiler_plate/firecrawl_boilerplate/pyproject.toml
[project]
name = "firecrawl-boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firecrawl-py>=2.8.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]
